AuditorHook breaks when faced with UnsavedRelationList.
Framework 3 has a bug where it promises to return array from
SS_List::map. RelationList implements SS_List - wrongly, because it
returns SS_Map instead of an array. This confused the auditor author to
treat the HasMany return value as SS_Map, which is incorrect.

The problem results in an "ERROR [User Error]: Uncaught Error: Call to a
member function toArray() on array" when for example publishing a page
with CanViewType set to OnlyTheseUsers for the first time.

This breach of contract in RelationList/DataList has been fixed in SS4.